Split CommandNamespace.kt into seperate files
