+++ src/main/kotlin/xyz/atrius/waystones/data/MultiReferencable.kt
@@ -1,0 +1,6 @@
+package xyz.atrius.waystones.data
+
+interface MultiReferencable {
+
+    val aliases: Array<out String>
+}
